Remove needless cast
currentTurn is already an int.

+++ src/rotp/model/galaxy/Galaxy.java
@@ -1,6 +1,6 @@
 /*
  * Copyright 2015-2020 Ray Fowler
- * Modifications Copyright 2023-2025 Ilya Zushinskiy
+ * Modifications Copyright 2023-2026 Ilya Zushinskiy
  * 
  * Licensed under the GNU General Public License, Version 3 (the "License");
  * you may not use this file except in compliance with the License.
@@ -293,8 +293,8 @@ public class Galaxy implements Base, Serializable {
        for (Empire emp: empires())
             emp.validate();
     }
-    public int numberTurns() { return (int) currentTurn; }
-    public int currentTurn() { return (int) currentTurn; }
+    public int numberTurns() { return currentTurn; }
+    public int currentTurn() { return currentTurn; }
 
     public Empire empire(int i)     {
         return (i < 0) || (i >= empires.length) ? null : empires[i];
